requirements/base.txt
amqp==5.3.1
anyio==4.8.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
billiard==4.2.1
black==23.1.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.11
cryptography==44.0.0
defusedxml==0.8.0rc2
diff-match-patch==20241021
dj-rest-auth==7.0.1
Django==5.0.1
django-admincharts==0.4.1
django-allauth==65.4.1
django-appconf==1.0.6
django-autoslug==1.9.8
django-celery-email==3.0.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.5
django-import-export==3.3.7
django-nine==0.2.7
django-phonenumber-field==7.0.2
django-taggit==3.1.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
drf-spectacular==0.22.1
drf-yasg==1.21.5
ecdsa==0.19.0
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
et_xmlfile==2.0.0
eventlet==0.39.1
factory_boy==3.3.3
Faker==18.3.1
flake8==6.0.0
flower==1.2.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gunicorn==20.1.0
httplib2==0.22.0
humanize==4.11.0
idna==3.10
importlib-metadata==4.13.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupPy==1.14
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.2.2
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
phonenumbers==8.13.7
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.10.1
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.2
redis==4.5.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.3
stripe==11.6.0
tablib==3.5.0
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
watchfiles==0.18.1
wcwidth==0.2.13
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.21.0
